data_ingestion_function/requirements.txt
google-cloud-firestore==2.16.0
googlemaps==4.10.0
requests==2.31.0